+++ src/main/java/de/skerkewitz/blubberblase/esc/component/TransformComponent.java
@@ -1,19 +1,19 @@
 package de.skerkewitz.blubberblase.esc.component;
 
-import de.skerkewitz.enora2d.common.Point2i;
+import de.skerkewitz.enora2d.common.Point2f;
 import de.skerkewitz.enora2d.core.ecs.component.Component;
 
 /**
  * The transform of an entity in world space.
  */
 public class TransformComponent implements Component {
-  public Point2i position;
+  public Point2f position;
 
   public TransformComponent() {
-    this.position = Point2i.ZERO;
+    this.position = Point2f.ZERO;
   }
 
-  public TransformComponent(Point2i position) {
+  public TransformComponent(Point2f position) {
     this.position = position;
   }
